requirements.txt
ag-ui-protocol==0.1.15
aiofile==3.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.89.0
anyio==4.13.0
argcomplete==3.6.3
asttokens==3.0.1
attrs==26.1.0
authlib==1.6.9
backoff==2.2.1
backports-tarfile==1.2.0
beartype==0.22.9
blinker==1.9.0
boto3==1.42.83
botocore==1.42.83
cachetools==7.0.5
caio==0.9.25
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cohere==5.21.1
cryptography==46.0.6
cyclopts==4.10.1
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
eval-type-backport==0.3.1
exceptiongroup==1.3.1
executing==2.2.1
faiss-cpu==1.13.2
fastavro==1.12.1
fastmcp==3.2.0
feedparser==6.0.12
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.3.0
genai-prices==0.0.56
gitdb==4.0.12
gitpython==3.1.46
google-auth==2.49.1
google-genai==1.70.0
googleapis-common-protos==1.74.0
graphiti==0.1.13
graphiti-core==0.28.2
griffelib==2.0.2
groq==1.1.2
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.9.0
idna==3.11
importlib-metadata==8.7.1
ipython==9.10.1
ipython-pygments-lexers==1.1.1
jaraco-classes==3.4.0
jaraco-context==6.1.2
jaraco-functools==4.4.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
jsonpatch==1.33
jsonpath-python==1.1.5
jsonpickle==4.1.1
jsonpointer==3.1.1
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
keyring==25.7.0
langchain==1.2.15
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.26
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.25
logfire==4.31.0
logfire-api==4.31.0
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mcp==1.27.0
mdurl==0.1.2
mistralai==2.3.0
more-itertools==11.0.1
multidict==6.7.1
mypy-extensions==1.1.0
narwhals==2.18.1
neo4j==6.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nexus-rpc==1.4.0
numpy==2.4.4
openai==2.30.0
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-httpx==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
orjson==3.11.8
ormsgpack==1.12.2
packaging==25.0
pandas==3.0.2
parso==0.8.6
pathable==0.5.0
pexpect==4.9.0
pillow==12.2.0
platformdirs==4.9.4
posthog==7.9.12
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
ptyprocess==0.7.0
pure-eval==0.2.3
py-key-value-aio==0.4.4
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-ai==1.77.0
pydantic-ai-slim==1.77.0
pydantic-core==2.41.5
pydantic-evals==1.77.0
pydantic-graph==1.77.0
pydantic-handlebars==0.1.0
pydantic-settings==2.13.1
pydeck==0.9.1
pygments==2.20.0
pyjwt==2.12.1
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.24
pytz==2026.1.post1
pyvis==0.3.2
pyyaml==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
s3transfer==0.16.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sqlalchemy==2.0.49
sse-starlette==3.3.4
stack-data==0.6.3
starlette==1.0.0
streamlit==1.56.0
temporalio==1.24.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
types-protobuf==7.34.1.20260403
types-requests==2.33.0.20260402
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
uncalled-for==0.2.0
urllib3==2.6.3
uuid-utils==0.14.1
uvicorn==0.43.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
wrapt==1.17.3
xai-sdk==1.11.0
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
